# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Textile ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-parties:
    get:
      operationId: listPurchaseParties
      tags: [purchaseParties]
      summary: List purchase parties
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasePartyList"
    post:
      operationId: createPurchaseParty
      tags: [purchaseParties]
      summary: Create purchase party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasePartyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseParty"
  /purchase-parties/{id}:
    get:
      operationId: getPurchaseParty
      tags: [purchaseParties]
      summary: Get purchase party by ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseParty"
    put:
      operationId: updatePurchaseParty
      tags: [purchaseParties]
      summary: Update purchase party
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasePartyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseParty"
    delete:
      operationId: deletePurchaseParty
      tags: [purchaseParties]
      summary: Delete purchase party
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Deleted

  # ── Sale Parties ─────────────────────────────────────────────────────────────
  /sale-parties:
    get:
      operationId: listSaleParties
      tags: [saleParties]
      summary: List sale parties
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalePartyList"
    post:
      operationId: createSaleParty
      tags: [saleParties]
      summary: Create sale party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalePartyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleParty"
  /sale-parties/{id}:
    get:
      operationId: getSaleParty
      tags: [saleParties]
      summary: Get sale party by ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleParty"
    put:
      operationId: updateSaleParty
      tags: [saleParties]
      summary: Update sale party
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalePartyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleParty"
    delete:
      operationId: deleteSaleParty
      tags: [saleParties]
      summary: Delete sale party
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Deleted

  # ── Shikanja ─────────────────────────────────────────────────────────────────
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
    offset:
      name: offset
      in: query
      schema:
        type: integer
...
  schemas:
...
    PurchaseParty:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameUrdu:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        openingCredit:
          type: ["string", "null"]
        openingDebit:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        shikanjaId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PurchasePartyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        nameUrdu:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        mobile:
          type: string
        openingCredit:
          type: number
        openingDebit:
          type: number
        type:
          type: string
          enum: [cash, credit]
        shikanjaId:
          type: integer
    PurchasePartyList:
      type: object
      required: [rows, total]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseParty"
        total:
          type: integer
    SaleParty:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameUrdu:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        creditLimit:
          type: ["string", "null"]
        openingCredit:
          type: ["string", "null"]
        openingDebit:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        shikanjaId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SalePartyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        nameUrdu:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        mobile:
          type: string
        creditLimit:
          type: number
        openingCredit:
          type: number
        openingDebit:
          type: number
        type:
          type: string
          enum: [cash, credit]
        shikanjaId:
          type: integer
    SalePartyList:
      type: object
      required: [rows, total]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/SaleParty"
        total:
          type: integer
...