# lib/api-zod/src/generated/types/purchaseParty.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Textile ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseParty {
  id: number;
  name: string;
  /** @nullable */
  nameUrdu?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  openingCredit?: string | null;
  /** @nullable */
  openingDebit?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  shikanjaId?: number | null;
  createdAt: string;
  updatedAt: string;
}