# lib/api-zod/src/generated/types/purchasePartyInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Textile ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { PurchasePartyInputType } from './purchasePartyInputType';

export interface PurchasePartyInput {
  name: string;
  nameUrdu?: string;
  address?: string;
  city?: string;
  phone?: string;
  mobile?: string;
  openingCredit?: number;
  openingDebit?: number;
  type?: PurchasePartyInputType;
  shikanjaId?: number;
}