# lib/api-zod/src/generated/types/salePartyInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Textile ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { SalePartyInputType } from './salePartyInputType';

export interface SalePartyInput {
  name: string;
  nameUrdu?: string;
  address?: string;
  city?: string;
  phone?: string;
  mobile?: string;
  creditLimit?: number;
  openingCredit?: number;
  openingDebit?: number;
  type?: SalePartyInputType;
  shikanjaId?: number;
}